idea-hn-47376584-show-hn--gitagent---an-open-standard-tha/requirements.txt
click>=8.1.0
gitpython>=3.1.0
pyyaml>=6.0
openai>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
